pyproject.toml
[tool.poetry]
name = "betsapi3"
version = "0.1.0"
description = "Unofficial BetsAPI wrapper for Python3"
repository = "https://github.com/tavkhelidzeluka/betsapi3-python"
readme = "README.md"
authors = ["tavkhelidzeluka <tavkhelidzeluka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
